Python-API-Service-Final-main/Python-API-Service-Final-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
PyJWT
Werkzeug